move throttle into stream directory
