+++ tokio/tests/time_throttle.rs
@@ -1,7 +1,7 @@
 #![warn(rust_2018_idioms)]
 #![cfg(feature = "full")]
 
-use tokio::stream::StreamExt;
+use tokio::stream::throttle;
 use tokio::time;
 use tokio_test::*;
 
@@ -11,9 +11,10 @@ use std::time::Duration;
 async fn usage() {
     time::pause();
 
-    let stream_repeat = futures::stream::repeat(());
-
-    let mut stream = task::spawn(stream_repeat.throttle(Duration::from_millis(100)));
+    let mut stream = task::spawn(throttle(
+        Duration::from_millis(100),
+        futures::stream::repeat(()),
+    ));
 
     assert_ready!(stream.poll_next());
     assert_pending!(stream.poll_next());
